# ci: consolidate and optimize jobs