--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -14,30 +14,6 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  dedupe:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Use Node.js 18.x
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18.x'
-      - name: Run yarn dedupe
-        run: yarn dedupe --check
-
-  format:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Use Node.js 18.x
-        uses: actions/setup-node@v3
-        with:
-          node-version: '18.x'
-      - name: Install dependencies
-        run: yarn install --immutable --immutable-cache
-      - name: Check formatting of project files
-        run: yarn format:diff
-
   lint:
     runs-on: ubuntu-latest
     steps:
@@ -46,8 +22,19 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: '18.x'
+      - uses: actions/cache@v3
+        id: cache
+        with:
+          path: |
+            node_modules
+            */**/node_modules
+          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
+      - name: Run yarn dedupe
+        run: yarn dedupe --check
       - name: Install dependencies
         run: yarn install --immutable --immutable-cache
+      - name: Check formatting of project files
+        run: yarn format:diff
       - name: Lint JavaScript files
         run: yarn lint
       - name: Lint Sass files