--- a/packages/react/icons/.gitignore
+++ b/packages/react/icons/.gitignore
@@ -1,2 +1,0 @@
-/index.esm.js
-/index.js