--- a/packages/react/icons/index.d.ts
+++ b/packages/react/icons/index.d.ts
@@ -1,0 +1,1 @@
+export * from '@carbon/icons-react';