--- a/packages/react/icons/index.js
+++ b/packages/react/icons/index.js
@@ -1,0 +1,1 @@
+module.exports = require('@carbon/icons-react');