--- a/packages/react/icons/package.json
+++ b/packages/react/icons/package.json
@@ -1,4 +1,0 @@
-{
-  "module": "index.esm.js",
-  "sideEffects": false
-}