--- a/packages/react/icons/src/index.js
+++ b/packages/react/icons/src/index.js
@@ -1,8 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2023
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-export * from '@carbon/icons-react';