Extension of https://github.com/carbon-design-system/carbon/pull/14560 which also addresses build and e2e

#### Changelog

**Changed**

- consolidated dedupe, format, lint, build, e2e

**Removed**

- build from avt and vrt

#### Testing / Reviewing

lint-build-e2e: 3m 27s vs 42s + 2m 57s + 2m 25s + 2m 25s
overall build time: 31 --> 25m