@dbrugger 2023-09-05
they still show as individual steps. no need to dig into logs to see which step failed:
<img width="321" alt="image" src="https://github.com/carbon-design-system/carbon/assets/10086178/44b68cce-67a1-4d93-8cde-8d432a154fbd">

Maybe no resource constraints are enforced, but am sure these improvements combined save ~30-50% compute. Carbon reducing Carbon emission :)