@dbrugger 2023-09-05
note the build itself can be improved as neither avt nor vrt really needs the full build. that is also part of one of these PRs, so that only what is needed is built instead of all packages.